+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Fix bulk importing of redirects (Fixes #154)
 - Fix social title not syncing for new entries
 - Fix snippet url preview bug (Fix #196 via @domstubbs)
+- Fix redirect uri case sensitivity issue with MySQL (Fixes #116)
 
 ## 3.6.0 - 2019-07-25
 ### Added